Build the final prompt via 'snprintf' instead of directly printing to stdout
It made it easier for debugging to have the entire prompt inside a variable.
Makes it possible to inspect its values.

Only print git stats if non zero
Changed how git stats is passed to the program. It is now expected to be 3 separate arguments
instead of just one that is split in C.
It is easier the split the argument via the shell.